requirements.txt
absl-py==2.1.0
apache-airflow==2.7.2
argcomplete==3.5.1
azure-core==1.32.0
azure-eventhub==5.13.0
azure-identity==1.19.0
blinker==1.9.0
boto3==1.35.95
botocore==1.35.95
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
msal==1.31.1
msal-extensions==1.2.0
numpy==2.2.2
packaging==24.1
pandas==2.2.3
pipx==1.7.1
platformdirs==4.3.6
portalocker==2.10.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32==308
requests==2.28.1
s3transfer==0.10.4
setuptools==75.6.0
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
userpath==1.9.2
Werkzeug==3.1.3
wheel==0.45.1
